atualização do mvp

+++ backend/src/users/users.service.ts
@@ -1,15 +1,19 @@
-import { HttpException, HttpStatus, Injectable, NotFoundException } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable, NotFoundException, BadRequestException } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { PrismaService } from 'src/prisma/prisma.service';
 import * as bcrypt from 'bcrypt';
 import { Bluetooth, Tag, User } from '@prisma/client';
+import { TagsService } from 'src/tags/tags.service';
 
 export const roundsOfHashing = 10;
 
 @Injectable()
 export class UsersService {
-  constructor(private readonly prisma: PrismaService) {}
+  constructor(
+    private readonly prisma: PrismaService,
+    private readonly Tags: TagsService = new TagsService(prisma)
+    ) {}
 
   async create(createUserDto: CreateUserDto): Promise<User> {
     const hashedPassword = await bcrypt.hash(
@@ -17,77 +21,103 @@ export class UsersService {
       roundsOfHashing,
     );
     
-    let user: User
+    const user = await this.prisma.user.create({
+      data: {
+        name: createUserDto.name,
+        registration: createUserDto.registration,
+        role: createUserDto.role,
+        password: hashedPassword
+      }
+    })
+    
+    let tag: Tag | undefined
+    let bluetooth: Bluetooth | undefined
+
+    if (createUserDto.bluetooth) {
+      bluetooth = await this.prisma.bluetooth.create({ 
+        data: { content: createUserDto.bluetooth }
+      }) 
+    }
+
+    if (createUserDto.tag) {
+      tag = await this.Tags.create({ content: createUserDto.tag, userId: user.id  })
+    }
+
+    return user;
+  }
+
+  async createAndLinkEnvironment(createUserDto: CreateUserDto, envId: number) {
+    const hashedPassword = await bcrypt.hash(
+      createUserDto.password,
+      roundsOfHashing,
+    );
+
+    const user = await this.prisma.user.create({
+      data: {
+        name: createUserDto.name,
+        registration: createUserDto.registration,
+        role: createUserDto.role,
+        password: hashedPassword,
+        adminEnvironment: createUserDto.role === 'ADMIN' ? { connect:{ id:envId } } : undefined,
+        frequenterEnvironment: createUserDto.role === 'FREQUENTER' ? { connect:{ id:envId } } : undefined
+      },
+    });
+
     let tag: Tag
     let bluetooth: Bluetooth
 
-    try {
-      if (createUserDto.tag) {
-        tag = await this.prisma.tag.create({ 
-          data: { content: createUserDto.tag }
-        }) 
-      }
-
-      if (createUserDto.bluetooth) {
-        bluetooth = await this.prisma.bluetooth.create({ 
-          data: { content: createUserDto.bluetooth }
-        }) 
-      }
-
-      user = await this.prisma.user.create({
-        data: {
-          ...createUserDto,
-          tag: createUserDto.tag ? { connect: { id: tag.id } } : undefined, // ou null
-          bluetooth: createUserDto.bluetooth ? { connect: { id: bluetooth.id } } : undefined, // ou null
-          password: hashedPassword
-        },
-      });
-
-      return user;
-    } catch (error) {
-      // Tratar erros específicos do Prisma, como por exemplo, violações de constraints ou índices.
-      // Aqui, estamos lançando um erro genérico, mas é possível criar erros mais específicos conforme a necessidade.
-      throw new Error('Erro ao criar usuário.');
+    if (createUserDto.bluetooth) {
+      bluetooth = await this.prisma.bluetooth.create({ 
+        data: { content: createUserDto.bluetooth }
+      }) 
     }
+
+    if (createUserDto.tag) {
+      tag = await this.Tags.create({ content: createUserDto.tag  })
+    }
+
+    return user;
   }
 
   async findAll() {
-    try {
-      return await this.prisma.user.findMany({
-        include: {
-          adminEnvironment: true,
-          frequenterEnvironment: true,
-          tag: true,
-          bluetooth: true
-        }
-      });
-    } catch (error) {
-      // Tratar erros específicos do Prisma, como por exemplo, conexões com o banco de dados.
-      throw new Error('Erro ao buscar usuários.');
-    }
+    return await this.prisma.user.findMany({
+      include: {
+        adminEnvironment: true,
+        frequenterEnvironment: true,
+        tag: true,
+        bluetooth: true
+      }
+    });
+  }
+
+  async findAllFrequenters() {
+    return await this.prisma.user.findMany({
+      where: { role: 'FREQUENTER' }
+    });
+  }
+
+  async findAllAdmins() {
+    return await this.prisma.user.findMany({
+      where: { role: 'ADMIN' }
+    });
   }
 
   async findOne(id: number) {
-    try {
-      const user = await this.prisma.user.findFirst({ 
-        where: { id },
-        include: {
-          adminEnvironment: true,
-          frequenterEnvironment: true,
-          tag: true,
-          bluetooth: true
-        }
-      });
-
-      if (!user) {
-        throw new NotFoundException('Usuário não encontrado.');
-      }
-
-      return user;
-    } catch (error) {
-      // Tratar erros específicos do Prisma, como por exemplo, conexões com o banco de dados.
-      throw new Error('Erro ao buscar usuário.');
+    if (!id) {
+      throw new BadRequestException('Invalid Input. ID must be sent.');
     }
+
+    const user = await this.prisma.user.findUniqueOrThrow({ 
+      where: { id },
+      include: {
+        adminEnvironment: true,
+        frequenterEnvironment: true,
+        tag: true,
+        bluetooth: true
+      }
+    });
+
+    return user;
   }
 
   async update(id: number, updateUserDto: UpdateUserDto) {
@@ -98,83 +128,39 @@ export class UsersService {
       );
     }
 
-    const user = await this.prisma.user.findUnique({
-      where: { id },
-      include: { tag: true, bluetooth: true },
-    });
-    
-    if (!user) {
-      throw new HttpException('User not found.', HttpStatus.NOT_FOUND);
-    }
+    const validFields = ['name', 'registration', 'password', 'role'];
+    const invalidFields = Object.keys(updateUserDto).filter(
+      field => !validFields.includes(field),
+    );
 
-    let tag: Tag | undefined;
-    let bluetooth: Bluetooth | undefined;
-
-    if (updateUserDto.tag) {
-      if (!user.tag) {
-        tag = await this.prisma.tag.create({
-          data: { content: updateUserDto.tag },
-        });
-      } else {
-        tag = await this.prisma.tag.update({
-          where: { id: user.tag.id },
-          data: { content: updateUserDto.tag },
-        });
-      }
-    }
-
-    if (!tag) {
-      throw new HttpException("Can't update tag.", HttpStatus.FORBIDDEN);
-    }
-
-
-    if (updateUserDto.bluetooth) {
-      if (!user.bluetooth) {
-        bluetooth = await this.prisma.bluetooth.create({
-          data: { content: updateUserDto.bluetooth },
-        });
-      } else {
-        bluetooth = await this.prisma.bluetooth.update({
-          where: { id: user.bluetooth.id },
-          data: { content: updateUserDto.bluetooth },
-        });
-      }
-    }
-
-    if (!bluetooth) {
-      throw new HttpException("Can't update bluetooth.", HttpStatus.FORBIDDEN);
-    }
+  if (invalidFields.length > 0) {
+    throw new BadRequestException(
+      `Invalid fields provided: ${invalidFields.join(', ')}`,
+    );
+  }
 
     const updatedUser = await this.prisma.user.update({
       data: {
-        ...updateUserDto,
-        tag: tag ? { connect: { id: tag.id } } : undefined,
-        bluetooth: bluetooth ? { connect: { id: bluetooth.id } } : undefined,
+        name: updateUserDto.name,
+        registration: updateUserDto.registration,
+        password: updateUserDto.password,
+        role: updateUserDto.role
       },
       where: { id }
     });
-
-    if (!updatedUser) {
-      throw new HttpException("Can't update user.", HttpStatus.FORBIDDEN);
-    }
 
     return updatedUser;
   }
 
   async remove(id: number) {
-    try {
-      const deletedUser = await this.prisma.user.delete({
-        where: { id }
-      });
-
-      if (!deletedUser) {
-        throw new NotFoundException('Usuário não encontrado.');
-      }
-
-      return deletedUser;
-    } catch (error) {
-      // Tratar erros específicos do Prisma, como por exemplo, conexões com o banco de dados.
-      throw new Error('Erro ao deletar usuário.');
+    if (isNaN(id)) {
+      throw new BadRequestException('Invalid input. ID must be a number.');
     }
+    
+    const deletedUser = await this.prisma.user.delete({
+      where: { id }
+    });
+
+    return deletedUser;
   }
 }
